+++ packages/sources/src/components/header/header.args.ts
@@ -10,6 +10,7 @@ export interface HeaderArgs {
     label: string;
     url: string;
   }[];
+  noMainMenu: boolean,
   useDropDownMenu: "always" | "never" | "auto";
   loginUrl: string;
   logoutUrl: string;
@@ -31,6 +32,12 @@ export const HeaderArgTypes: ArgTypes<HeaderArgs> = {
     },
   },
   mainMenu: noControl,
+  noMainMenu: {
+    name: "Geen menu",
+    control: {
+      type: "boolean",
+    },
+  },
   useDropDownMenu: {
     name: "Dropdown menu",
     options: ["always", "never", "auto"],
@@ -76,7 +83,7 @@ export function headerArgsMapper(a: HeaderArgs): Header {
     logo: a.logo,
     subLogo: a.subLogo,
     showSubLogo: a.showSubLogo,
-    mainMenu: a.mainMenu,
+    mainMenu: a.noMainMenu ? undefined : a.mainMenu,
     useDropDownMenu: a.useDropDownMenu,
     loginUrl: a.loginUrl,
     logoutUrl: a.logoutUrl,